+++ internal/engine/compiler/compiler_test.go
@@ -273,7 +273,6 @@ type compilerImpl interface {
 	getOnStackPointerCeilDeterminedCallBack() func(uint64)
 	setStackPointerCeil(uint64)
 	compileReleaseRegisterToStack(loc *runtimeValueLocation)
-	runtimeValueLocationStack() *runtimeValueLocationStack
 	setRuntimeValueLocationStack(*runtimeValueLocationStack)
 	compileEnsureOnRegister(loc *runtimeValueLocation) error
 	compileModuleContextInitialization() error
